refactor!: move factory_root_keys mod code to shared
Also update README completed commands.
